+++ src/constant/apiKey.js
@@ -1,0 +1,1 @@
+export const DOCUMENT_KEY = "/documents";
